Add files via upload
1. class Spaceship
class: This keyword tells Python we’re defining a new class, which is a blueprint for creating objects. Here, our blueprint is for a Spaceship.
Spaceship: The name of the class, which we use as the model for creating spaceship objects.

2. def __init__(self, name, speed, fuel)
def: This keyword is short for "define," and it tells Python we’re creating a function or method. A function is a reusable block of code that performs a specific task.
__init__: This is a special method called a constructor. It automatically runs when we create a new instance (or object) of the class.
Parameters (self, name, speed, fuel):
self: Refers to the specific instance (object) of the class that’s being created. By convention, we use self to represent the object.
name, speed, fuel: These are values that we pass in when creating a new spaceship. They’re used to set the initial attributes for our object.
Inside the __init__ Method
python
Copy code
self.name = name
self.speed = speed
self.fuel = fuel
self.name = name: This assigns the value of name passed to the object’s attribute name.
The same pattern goes for self.speed and self.fuel, which store the spaceship’s speed and fuel.
So, when we create a spaceship with Spaceship("Apollo", 500, 100), the name becomes "Apollo", speed is 500, and fuel is 100.

3. def fly(self, distance)
This is another method defined with def.
fly: This is a custom method we created to simulate the spaceship flying a certain distance. When fly is called, it checks if there’s enough fuel and updates the spaceship’s fuel level accordingly.

4. fuel_needed = distance * 0.1
This line calculates how much fuel is needed for the trip.
Here, distance * 0.1 assumes that for every unit of distance, the spaceship consumes 0.1 units of fuel.
The result is stored in fuel_needed, which will be used in the next part to determine if the spaceship can fly the specified distance.

5. if-else Statement
python
Copy code
if self.fuel >= fuel_needed:
    self.fuel -= fuel_needed
    print(f"{self.name} flew {distance} units!")
else:
    print(f"{self.name} doesn't have enough fuel to fly!")
if self.fuel >= fuel_needed: This checks if the spaceship’s current fuel (self.fuel) is greater than or equal to the required fuel (fuel_needed).
self.fuel -= fuel_needed: If the spaceship has enough fuel, this line subtracts fuel_needed from self.fuel to represent using up the fuel for the trip.
print(f"{self.name} flew {distance} units!"): This prints a message that the spaceship has flown the distance.
else: If the spaceship doesn’t have enough fuel, the code inside else executes.
print(f"{self.name} doesn't have enough fuel to fly!"): This message alerts that the spaceship can’t complete the trip due to low fuel.

6. Creating and Using the Spaceship Instance
python
Copy code
spaceship = Spaceship("Apollo", 500, 100)
spaceship.fly(200)
spaceship = Spaceship("Apollo", 500, 100): This line creates a new instance (object) of the Spaceship class, named spaceship.
Here, "Apollo" is the name, 500 is the speed, and 100 is the fuel.
spaceship.fly(200): Calls the fly method on our spaceship instance, telling it to attempt to fly a distance of 200.
This setup now allows you to create multiple Spaceship objects with different names, speeds, and fuel levels, and use the fly method to simulate their journeys!

Let me know if this makes sense or if you’d like more detail on any part.